Migrate from Ava to Jest
AVA is a highly opinionated, minimalist testing framework, so it misses some features like test grouping and code coverage, and
it does not use globals, which would normally be a good thing, but the API also increases verbosity. At the same time Jest is more
popular, and it has more readable assertions. It is true that it does not support custom failure messages, but it also prints
the failing code lines, which is enough considering the code should also be readable.