Let the project specify the Cypress command to execute
Also added some tests so we can test various invocations:

* with and without docker,
* with and without a serve command also executed.

For this to work realiably, I decided not to test against [a web app that is not under our control](https://example.cypress.io/),
which is also unnecessarily slower. Therefore, I added a fake web app with a fake development web server.

Note: originally I added a test case that used `--serveCmd` but docker was not enabled. For some reason it hang on GitHub,
but I could not reproduce locally. I did not investigate this case further but removed the test case.